refactor(whispering): dissolve AppLayout into a stable AppRuntime owner
Fixes #2035: the version-update notification (and the setup reminder) fired
again every time the window was resized across the 767px nav breakpoint.

Root cause: AppLayout was rendered in both arms of the responsive {#if} in
the (app) layout, so crossing the breakpoint destroyed one instance and
mounted a fresh one, re-running its onMount -> checkForUpdates() each time.
report.info has no toast key, so every check stacked another toast.

The real problem was ownership: app-lifetime bootstrap was attached to a
view component the responsive tree legitimately remounts. Split the two
concerns:

- AppRuntime.svelte: headless, mounted once at the session root, outside the
  {#if}. Owns all bootstrap (merged from both +layout and AppLayout). Mounts
  exactly once per launch, so once-per-launch is structural, not a guard.
- ContentShell.svelte: the recording-button toggle + page content. Pure view;
  free to remount on a breakpoint change.
- GlobalDialogs.svelte: the three singleton dialogs, mounted once (also stops
  a dialog tearing down mid-resize).
- +layout.svelte: thin session root that renders the three plus the chrome.

AppLayout is deleted; its name described a layout that also booted the app.
